third_party/WebKit/LayoutTests/shadow-dom/v1-slots-in-document-tree.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script src='../fast/dom/shadow/resources/shadow-dom.js'></script>
<div id='wrapper'>
  <slot id='slot1'></slot>
  <slot id='slot2'>
    <div id='child1'></div>
  </slot>
  <slot id='slot3'>
    <div id='child2'></div>
    <slot id='slot4'>
      <div id='child3'></div>
      <div id='child4'></div>
    </slot>
  </slot>
</div>
<script>
'use strict';
removeWhiteSpaceOnlyTextNodes(wrapper);

test(() => {
  assert_equals(child1.assignedSlot, null);
  assert_equals(child2.assignedSlot, null);
  assert_equals(child3.assignedSlot, null);
  assert_equals(child4.assignedSlot, null);
}, "assignedSlot");

test(() => {
  assert_array_equals(slot1.getAssignedNodes(), []);
  assert_array_equals(slot2.getAssignedNodes(), []);
  assert_array_equals(slot3.getAssignedNodes(), []);
  assert_array_equals(slot4.getAssignedNodes(), []);
}, "getAssignedNodes");

test(() => {
  assert_array_equals(slot1.getAssignedNodes({flatten: true}), []);
  assert_array_equals(slot2.getAssignedNodes({flatten: true}), [child1]);
  assert_array_equals(slot3.getAssignedNodes({flatten: true}), [child2, child3, child4]);
  assert_array_equals(slot4.getAssignedNodes({flatten: true}), [child3, child4]);
}, "getAssignedNodes({flatten: true})");
</script>
